pyproject.toml
[project]
name = "internship-practice"
version = "0.1.0"
description = ""
authors = [
    {name = "kamleshcode",email = "kamleshit21@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13,<3.15"
dependencies = [
    "pandas (>=1.4.0,<3)",
    "flask (>=3.1.2,<4.0.0)",
    "flasgger (>=0.9.7.1,<0.10.0.0)",
    "fastapi (>=0.128.2,<0.129.0)",
    "uvicorn (>=0.40.0,<0.41.0)",
    "pydantic (>=2.12.5,<3.0.0)",
    "sqlalchemy (>=2.0.46,<3.0.0)",
    "pyodbc (>=5.3.0,<6.0.0)",
    "dotenv (>=0.9.9,<0.10.0)",
    "sqlalchemy-utils (>=0.42.1,<0.43.0)",
    "requests (>=2.32.5,<3.0.0)",
    "beautifulsoup4 (>=4.14.3,<5.0.0)",
    "selenium (>=4.40.0,<5.0.0)",
    "scikit-learn (>=1.8.0,<2.0.0)",
    "seaborn (>=0.13.2,<0.14.0)",
    "imblearn (>=0.0,<0.1)",
    "category-encoders (>=2.9.0,<3.0.0)",
    "plotly (>=6.6.0,<7.0.0)",
    "mlxtend (>=0.24.0,<0.25.0)",
    "tensorflow (>=2.21.0,<3.0.0)",
    "langchain (>=1.2.13,<2.0.0)",
    "langchain-community (>=0.4.1,<0.5.0)",
    "langchain-core (>=1.2.20,<2.0.0)",
    "langchain-ollama (>=1.0.1,<2.0.0)",
    "langchain-groq (>=1.1.2,<2.0.0)",
    "tavily-python (>=0.7.23,<0.8.0)",
    "langchain-tavily (>=0.2.17,<0.3.0)",
    "streamlit (>=1.55.0,<2.0.0)",
    "pypdf (>=6.9.2,<7.0.0)",
    "langchain-mongodb (>=0.11.0,<0.12.0)",
    "langsmith (>=0.7.22,<0.8.0)",
    "pymongo (==4.12)",
    "chromadb (>=1.5.7,<2.0.0)",
    "langchain-mistralai (>=1.1.2,<2.0.0)",
    "pymupdf (>=1.27.2.2,<2.0.0.0)",
    "langgraph-checkpoint-mongodb (>=0.3.1,<0.4.0)",
    "langchain-chroma (>=1.1.0,<2.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
